requirements.txt
streamlit
transformers
torch
Pillow
numpy
scikit-learn
plotly
